refactor: adjust imports and fixe require cycle
- removed extra `consulHelper` as we have a separate folder
  for consul helpers
  - moved contents to its index file
- adjusted comments for ts-error ignores that need to be by line
