+++ src/helpers/index.js
@@ -6,16 +6,17 @@ export * from './bookmarkHelper';
 export * from './calendarHelper';
 export * from './cameraHelper';
 export * from './colorHelper';
-export * from './consulHelper';
 export * from './dateTimeHelper';
 export * from './deleteArrayItem';
 export * from './encounterHelper';
 export * from './fileSizeHelper';
 export * from './formatHelper';
+export * from './genericTypeHelper';
 export * from './graphqlHelper';
 export * from './headerHelper';
 export * from './htmlViewHelper';
 export * from './imageHelper';
+export * from './initializationHelper';
 export * from './linkHelper';
 export * from './mapHelper';
 export * from './matomoHelper';
